backend/cloud_functions/universal_nlp_enricher/requirements.txt
google-cloud-bigquery>=3.12.0
google-cloud-language>=2.11.1
google-auth>=2.23.0
google-api-python-client>=2.100.0
functions-framework>=3.4.0

